[onert] Move feature reader and view
Move feature reader and view from include/util to src/exec
