+++ runtime/onert/core/src/exec/Source.h
@@ -17,13 +17,13 @@
 #ifndef __ONERT_EXEC_SOURCE_H__
 #define __ONERT_EXEC_SOURCE_H__
 
-#include <cassert>
+#include "feature/nchw/Reader.h"
+#include "feature/nchw/View.h"
+#include "feature/nhwc/Reader.h"
+#include "feature/nhwc/View.h"
 
+#include <cassert>
 #include <memory>
-#include "util/feature/nchw/Reader.h"
-#include "util/feature/nchw/View.h"
-#include "util/feature/nhwc/Reader.h"
-#include "util/feature/nhwc/View.h"
 #include "util/Utils.h"
 #include <misc/feature/IndexIterator.h>
 #include <ir/Layout.h>
@@ -132,8 +132,8 @@ protected:
 
           if (_io_layout == ir::Layout::NCHW)
           {
-            const util::feature::nchw::Reader<T> from(shape, _input_buffer, _input_size);
-            util::feature::nhwc::View<T> into(&tensor);
+            const exec::feature::nchw::Reader<T> from(shape, _input_buffer, _input_size);
+            exec::feature::nhwc::View<T> into(&tensor);
             ::nnfw::misc::feature::iterate(shape)
                 << [&](uint32_t batch, uint32_t ch, uint32_t row, uint32_t col) {
                      const auto value = from.at(batch, ch, row, col);
@@ -142,8 +142,8 @@ protected:
           }
           else if (_io_layout == ir::Layout::NHWC)
           {
-            const util::feature::nhwc::Reader<T> from(shape, _input_buffer, _input_size);
-            util::feature::nchw::View<T> into(&tensor);
+            const exec::feature::nhwc::Reader<T> from(shape, _input_buffer, _input_size);
+            exec::feature::nchw::View<T> into(&tensor);
             ::nnfw::misc::feature::iterate(shape)
                 << [&](uint32_t batch, uint32_t ch, uint32_t row, uint32_t col) {
                      const auto value = from.at(batch, row, col, ch);
