+++ runtime/onert/core/src/exec/feature/nchw/View.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 Samsung Electronics Co., Ltd. All Rights Reserved
+ * Copyright (c) 2018 Samsung Electronics Co., Ltd. All Rights Reserved
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,34 +14,35 @@
  * limitations under the License.
  */
 
-#ifndef __ONERT_UTIL_FEATURE_NCHW_READER_H__
-#define __ONERT_UTIL_FEATURE_NCHW_READER_H__
+#ifndef __ONERT_EXEC_FEATURE_NCHW_VIEW_H__
+#define __ONERT_EXEC_FEATURE_NCHW_VIEW_H__
+
+#include "../Reader.h"
+
+#include "misc/feature/Shape.h"
+#include "backend/ITensor.h"
+#include "util/logging.h"
 
 #include <cassert>
 
-#include "backend/ITensor.h"
-#include "misc/feature/Shape.h"
-#include "util/feature/Reader.h"
-
 namespace onert
 {
-namespace util
+namespace exec
 {
 namespace feature
 {
 namespace nchw
 {
 
-template <typename T> class Reader final : public feature::Reader<T>
+template <typename T> class View final : public feature::Reader<T>
 {
 public:
   // Construct for buffer of model inputs
-  Reader(const ::nnfw::misc::feature::Shape &shape, const T *ptr, size_t len)
-      : _shape{shape}, _ptr{reinterpret_cast<const uint8_t *>(ptr)}, _len{len}
+  View(const ::nnfw::misc::feature::Shape &shape, T *ptr, size_t len)
+      : _shape{shape}, _ptr{reinterpret_cast<uint8_t *>(ptr)}, _len{len}
   {
     assert(shape.N * shape.C * shape.H * shape.W * sizeof(T) == len);
 
-    // No padding
     _strides.W = sizeof(T);
     _strides.H = shape.W * sizeof(T);
     _strides.C = shape.W * shape.H * sizeof(T);
@@ -49,7 +50,7 @@ public:
   }
 
   // Construct for backend tensor
-  Reader(backend::ITensor *tensor)
+  View(::onert::backend::ITensor *tensor)
       : _ptr{tensor->buffer() + tensor->calcOffset({0, 0, 0, 0})}, _len{tensor->total_size()}
   {
     assert(tensor->layout() == ir::Layout::NCHW);
@@ -71,7 +72,7 @@ public:
   {
     const auto offset = feature_index_to_byte_offset(0, ch, row, col);
 
-    const T *ptr = reinterpret_cast<const T *>(_ptr + offset);
+    T *ptr = reinterpret_cast<T *>(_ptr + offset);
 
     return *ptr;
   }
@@ -79,7 +80,25 @@ public:
   {
     const auto offset = feature_index_to_byte_offset(batch, ch, row, col);
 
-    const T *ptr = reinterpret_cast<const T *>(_ptr + offset);
+    T *ptr = reinterpret_cast<T *>(_ptr + offset);
+
+    return *ptr;
+  }
+
+public:
+  T &at(uint32_t ch, uint32_t row, uint32_t col)
+  {
+    const auto offset = feature_index_to_byte_offset(0, ch, row, col);
+
+    T *ptr = reinterpret_cast<T *>(_ptr + offset);
+
+    return *ptr;
+  }
+  T &at(uint32_t batch, uint32_t ch, uint32_t row, uint32_t col)
+  {
+    const auto offset = feature_index_to_byte_offset(batch, ch, row, col);
+
+    T *ptr = reinterpret_cast<T *>(_ptr + offset);
 
     return *ptr;
   }
@@ -106,13 +125,13 @@ private:
   nnfw::misc::feature::Shape _shape;
   using Strides = nnfw::misc::feature::Shape;
   Strides _strides;
-  const uint8_t *_ptr;
+  uint8_t *_ptr;
   size_t _len;
 };
 
 } // namespace nchw
 } // namespace feature
-} // namespace util
+} // namespace exec
 } // namespace onert
 
-#endif // __ONERT_UTIL_FEATURE_NCHW_READER_H__
+#endif // __ONERT_EXEC_FEATURE_NCHW_VIEW_H__
